Gestion des erreurs HTTP request 400, 500

+++ BACKEND/node-film-api/src/routes/deleteFilm.js
@@ -3,8 +3,13 @@ const { Film } = require('../db/sequelize')
 module.exports = (app) => {
     app.delete('/api/films/:id', (req, res) => {
         Film.findByPk(req.params.id).then(film => {
-            const filmDeleted = pokemon;
-            Film.destroy({
+            if (film === null) {
+                const message = `Le film demandé n'existe pas. Réessayez avec un autre identifiant`;
+                return res.status(404).json({ messsage })
+            }
+
+            const filmDeleted = film;
+            return Film.destroy({
                 where: { id: film.id }
             })
                 .then(_ => {
@@ -12,5 +17,9 @@ module.exports = (app) => {
                     res.json({ message, data: filmDeleted })
                 })
         })
+            .catch(error => {
+                const message = `Le film n'a pas pu être supprimé. Réessayez dans quelques instants.`
+                res.status(500).json({ message, data: error })
+            })
     })
 }
